refactor: update About page to redirect to external TimeHarbor website and archive old content
